hispanshield-desktop-forensic/app/requirements.txt
PySide6
pefile
requests
reportlab
python-magic-bin; platform_system == 'Windows'
python-magic; platform_system != 'Windows'
yara-python
filetype
